fix(studio): handle stale staging branch in save/feedback endpoints
Extract ensureStagingBranch helper that:
- verifies a KV-cached branch still exists on GitHub before reusing it
  (branch may be deleted after a PR is merged)
- gracefully handles 422 "Reference already exists" when creating a branch
  that was created same-day after a publish cleared KV
