+++ src/routes/api/studio/feedback/+server.ts
@@ -1,5 +1,6 @@
 import type { RequestHandler } from './$types';
 import { Octokit } from '@octokit/rest';
+import { ensureStagingBranch, stagingKey, type StagingState } from '$lib/studio/staging';
 
 export const prerender = false;
 
@@ -11,15 +12,6 @@ type SubmissionType = 'cfp' | 'cfe';
 
 function feedbackPath(type: SubmissionType): string {
 	return `content/2026/feedback/${type}.json`;
-}
-
-function stagingKey(handle: string): string {
-	return `studio_staging:${handle}`;
-}
-
-interface StagingState {
-	branch: string;
-	files: string[];
 }
 
 interface FeedbackEntry {
@@ -77,27 +69,7 @@ export const POST: RequestHandler = async ({ request, locals, platform }) => {
 	const repoPath = feedbackPath(type as SubmissionType);
 
 	try {
-		// Load or create staging branch
-		const existing = (await kv.get(stagingKey(handle), 'json')) as StagingState | null;
-
-		let branchName: string;
-		if (existing?.branch) {
-			branchName = existing.branch;
-		} else {
-			const date = new Date().toISOString().slice(0, 10);
-			branchName = `studio/${handle}/${date}`;
-			const { data: ref } = await octokit.git.getRef({
-				owner: OWNER,
-				repo: REPO,
-				ref: `heads/${BASE_BRANCH}`
-			});
-			await octokit.git.createRef({
-				owner: OWNER,
-				repo: REPO,
-				ref: `refs/heads/${branchName}`,
-				sha: ref.object.sha
-			});
-		}
+		const { branchName, existing } = await ensureStagingBranch(octokit, kv, handle);
 
 		// Fetch current feedback JSON from the staging branch (fall back to master)
 		let currentFeedback: Record<string, FeedbackEntry> = {};
